cast last_auth_time from time_t to long int
git-svn-id: https://svn.code.sf.net/p/vpopmail/code/trunk@292 8f0cf6e2-dddb-4e5e-b901-7f5e1d474549

+++ vpopmail/vpopmaild.c
@@ -1572,7 +1572,7 @@ int get_lastauth()
   strncpy(WriteBuf,RET_OK_MORE,sizeof(WriteBuf));
   wait_write();
 
-  snprintf(WriteBuf, sizeof(WriteBuf), "%ld" RET_CRLF, last_auth_time);
+  snprintf(WriteBuf, sizeof(WriteBuf), "%ld" RET_CRLF, (long int) last_auth_time);
   return(0);
 }
